+++ satella/coding/structures/dictionaries/__init__.py
@@ -1,0 +1,3 @@
+from .objects import DirtyDict, DictionaryView, KeyAwareDefaultDict, TwoWayDictionary
+from .dict_object import apply_dict_object, DictObject
+from .expiring import ExpiringEntryDict, SelfCleaningDefaultDict