+++ CHANGELOG.md
@@ -4,4 +4,4 @@
 * added __slots__ to `ReprableMixin` and `Multirun`
 * added `read_in_file` and `write_to_file`
 * optimized `TimeBasedHeap`
-* added `TimeBasedSetHeap`
+* added `TimeBasedSetHeap` and `SelfCleaningDefaultDict`